- fixed wrong IgnoreProperties for certain effects - fixed some skill positions got blocked by shields - fixed material zone size calculation (e.g. fortress shields would not trigger when passing through from certain points) - teleport/movement skills now work over hills and also consider the ground when moving downhill (you also cannot fear someone over a cliff anymore) - fixed DespawnableNode instances bitset was not thread safe - necessary refactoring due to the new AbstractMaterialSkillActor class and other changes - improved IgnoreProperties to get rid of Race.ANY (which should not have been introduced in the first place) and added constants

+++ game-server/src/com/aionemu/gameserver/model/templates/materials/MaterialSkill.java
@@ -21,7 +21,7 @@ public class MaterialSkill {
 	private List<MaterialActCondition> conditions;
 
 	@XmlAttribute(required = true)
-	private float frequency;
+	private int frequency;
 
 	@XmlAttribute
 	private MaterialTarget target;
@@ -36,7 +36,7 @@ public class MaterialSkill {
 		return conditions == null ? Collections.emptyList() : conditions;
 	}
 
-	public float getFrequency() {
+	public int getFrequency() {
 		return frequency;
 	}
